be/socketController: phase2 / art vote on, off

See merge request s07-blockchain-nft-sub2/S07P22A509!59

+++ server/src/main/java/com/classic/imteller/api/service/ArtService.java
@@ -82,6 +82,7 @@ public class ArtService {
                     .owner(user)
                     .ownerNickname(user.getNickname())
                     .url("https://imtellercard.s3.ap-northeast-2.amazonaws.com/" + imgPath)
+                    .isVote(false)
                     .title(paintSaveReqDto.getPaintTitle())
                     .description(paintSaveReqDto.getDescription()).build();
             artRepository.save(art);
@@ -116,8 +117,48 @@ public class ArtService {
     }
 
     @Transactional
-    public void deletePaint(Long id) {
-        artRepository.deleteById(id);
+    public boolean deletePaint(Long id, String email) {
+        Art art = artRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.POSTS_NOT_FOUND));
+        if (art.getOwner().getEmail().equals(email)) {
+            artRepository.deleteById(id);
+            return true;
+        }
+        else return false;
+    }
+
+    @Transactional
+    public boolean onVote(Long id, String email) {
+        System.out.println(id + " " + email);
+        Art art = artRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.POSTS_NOT_FOUND));
+        if (art.getOwner().getEmail().equals(email)) {
+            art.updateIsVote(true);
+            artRepository.save(art);
+            return true;
+        }
+        else return false;
+    }
+
+    @Transactional
+    public boolean offVote(Long id, String email) {
+        Art art = artRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.POSTS_NOT_FOUND));
+        if (art.getOwner().getEmail().equals(email)) {
+            art.updateIsVote(false);
+            artRepository.save(art);
+            return true;
+        }
+        else return false;
+    }
+
+    @Transactional
+    public void insertTokenId(Long artId, Long tokenId) {
+        Art art = artRepository.findById(artId).orElseThrow(() -> new CustomException(ErrorCode.POSTS_NOT_FOUND));
+        art.insertNft(tokenId);
+    }
+    @Transactional
+    public void editOwner(String owner, Long tokenId) {
+        Art art = artRepository.findByTokenId(tokenId).orElseThrow(() -> new CustomException(ErrorCode.POSTS_NOT_FOUND));
+        User user = userRepository.findByNickname(owner).orElseThrow(() -> new CustomException(ErrorCode.POSTS_NOT_FOUND));
+        art.updateOwner(user, user.getNickname());
     }
 
     @Transactional
